+++ objectbox_test/test/sync_test.dart
@@ -37,9 +37,13 @@ void main() {
     expect(waitUntil(() => client.state() == SyncState.loggedIn), isTrue);
   }
 
-  // lambda to easily create clients in the test below
+  // lambda to easily create clients in the tests below
+  SyncClient createAuthenticatedClient(
+          Store s, List<SyncCredentials> credentials) =>
+      Sync.clientMultiCredentials(s, 'ws://127.0.0.1:$serverPort', credentials);
+
   SyncClient createClient(Store s) =>
-      Sync.client(s, 'ws://127.0.0.1:$serverPort', SyncCredentials.none());
+      createAuthenticatedClient(s, [SyncCredentials.none()]);
 
   // lambda to easily create clients in the test below
   SyncClient loggedInClient(Store s) {
@@ -57,10 +61,18 @@ void main() {
   });
 
   test('Sync.clientMulti throws if empty URL list', () {
+    // Note: this test works with a library that does not have the Sync
+    // feature, because the URLs are checked before checking for the feature.
     expect(
         () => Sync.clientMultiUrls(store, [], SyncCredentials.none()),
         throwsA(isArgumentError.having((e) => e.message, 'message',
-            contains('must contain at least one server URL'))));
+            contains('Provide at least one server URL'))));
+
+    expect(
+        () => Sync.clientMultiCredentialsMultiUrls(
+            store, [], [SyncCredentials.none()]),
+        throwsA(isArgumentError.having((e) => e.message, 'message',
+            contains('Provide at least one server URL'))));
   });
 
   test('SyncCredentials string encoding', () {
@@ -114,6 +126,18 @@ void main() {
       expect(store.syncClient(), isNull);
     });
 
+    test('Sync.clientMulti throws if empty credential list', () {
+      expect(
+          () => Sync.clientMultiCredentials(store, 'test-url', []),
+          throwsA(isArgumentError.having((e) => e.message, 'message',
+              contains('Provide at least one credential'))));
+
+      expect(
+          () => Sync.clientMultiCredentialsMultiUrls(store, ['test-url'], []),
+          throwsA(isArgumentError.having((e) => e.message, 'message',
+              contains('Provide at least one credential'))));
+    });
+
     test('SyncClient is closed when a store is closed', () {
       final client = createClient(env2.store);
       env2.closeAndDelete();
@@ -150,13 +174,30 @@ void main() {
       expect(() => c.cancelUpdates(), error);
       expect(() => c.requestUpdates(subscribeForFuturePushes: true), error);
       expect(() => c.outgoingMessageCount(), error);
+
       expect(() => c.setCredentials(SyncCredentials.none()), error);
+      expect(
+          () => c.setCredentials(SyncCredentials.sharedSecretString('secret')),
+          error);
+      expect(
+          () => c
+              .setCredentials(SyncCredentials.userAndPassword('obx', 'secret')),
+          error);
+
+      expect(
+          () => c.setMultipleCredentials([
+                SyncCredentials.sharedSecretString('secret'),
+                SyncCredentials.userAndPassword('obx', 'secret')
+              ]),
+          error);
+
       expect(() => c.setRequestUpdatesMode(SyncRequestUpdatesMode.auto), error);
     });
 
     test('SyncClient simple coverage (no server available)', () {
       SyncClient c = createClient(store);
       expect(c.isClosed(), isFalse);
+
       c.setCredentials(SyncCredentials.none());
       c.setCredentials(SyncCredentials.googleAuthString('secret'));
       c.setCredentials(SyncCredentials.sharedSecretString('secret'));
@@ -165,6 +206,11 @@ void main() {
       c.setCredentials(SyncCredentials.sharedSecretUint8List(
           Uint8List.fromList([13, 0, 25])));
       c.setCredentials(SyncCredentials.userAndPassword('obx', 'secret'));
+      c.setCredentials(SyncCredentials.jwtIdToken('id-token'));
+      c.setCredentials(SyncCredentials.jwtAccessToken('access-token'));
+      c.setCredentials(SyncCredentials.jwtRefreshToken('refresh-token'));
+      c.setCredentials(SyncCredentials.jwtCustomToken('custom-token'));
+
       c.setCredentials(SyncCredentials.none());
       c.setRequestUpdatesMode(SyncRequestUpdatesMode.manual);
       c.start();
@@ -176,7 +222,33 @@ void main() {
       expect(c.state(), equals(SyncState.stopped));
     });
 
-    group('Sync tests with server', () {
+    test('SyncClient setMultipleCredentials', () {
+      SyncClient c = createClient(store);
+
+      expect(
+          () => c.setMultipleCredentials([]),
+          throwsA(isA<ArgumentError>()
+              .having((e) => e.name, "name", "credentials")));
+
+      // none() not supported
+      expect(
+          () => c.setMultipleCredentials([SyncCredentials.none()]),
+          throwsA(isA<ArgumentError>()
+              .having((e) => e.name, "name", "credentials")));
+
+      // Not throwing in Dart for any supported type
+      c.setMultipleCredentials([
+        SyncCredentials.googleAuthString('secret'),
+        SyncCredentials.sharedSecretString('secret'),
+        SyncCredentials.userAndPassword('obx', 'secret'),
+        SyncCredentials.jwtIdToken('id-token'),
+        SyncCredentials.jwtAccessToken('access-token'),
+        SyncCredentials.jwtRefreshToken('refresh-token'),
+        SyncCredentials.jwtCustomToken('custom-token')
+      ]);
+    });
+
+    group('Server tests using sync-server in PATH', () {
       late SyncServer server;
 
       setUp(() async {
@@ -389,6 +461,57 @@ void main() {
         skip: SyncServer.isAvailable()
             ? null
             : 'sync-server executable is not available in PATH - tests requiring it are skipped');
+
+    group('Server tests expecting running Sync server', () {
+      final String testJwtToken = "INSERT_VALID_JWT";
+      final String testInvalidJwtToken =
+          "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiJzeW5jLXNlcnZlciIsImlzcyI6Im9iamVjdGJveC1hdXRoIiwiZXhwIjoxNzM4MjE1NjAwLCJpYXQiOjE3MzgyMTc0MDN9.3auqtgaSEqpFqXhuCyoDM-LbfTOIEGGF6X0AjCcykJ2Nv1WN6LaVbuMDjMf-tKSLyeqFkzQbIckP4FvLHh7wQJ6rafDiT4H2pb6xhouU1QH3szK2S_7VDl_4BhxRbW5pEUt9086HXaVFHEZVS0417pxomlPHxrc1n4Z_A4QxZM5_xh5xcHV8PiGgXWb6_2basjBj5z6POTrazRs67IOQ-ob6ROIsOUGu3om6b8i0h_QSMmeJbujfr2EZqhYWTKijeyidbjRWZ97NFxtGRYN_jPOvy-T3gANXs2a32Er8XvgZTjr_-O8tl_1fHPo2kDE-UCNdwUfBQFhTokDUdJ81bg";
+
+      /// NOTE Unlike the other tests, this test assumes a Sync server is
+      /// already running on [serverPort] and has JWT auth configured.
+      /// Then obtain a JWT and insert it in [testJwtToken] above.
+      ///
+      /// Background: Sync server needs to run in a supported environment to
+      /// enable JWT authentication. So this test can not interact with a
+      /// sync-server binary like the other tests.
+      test('Auth with JSON Web Token (JWT)', () async {
+        // Note: the objectbox project covers all cases, this test just
+        // ensures the Dart parts work as expected.
+
+        expect(testJwtToken, isNot("INSERT_VALID_JWT"),
+            reason:
+                "Paste a valid JWT into testJwtToken before running this test");
+
+        // Using an already running server, at least check it's available
+        await SyncServer.onlineAt(serverPort);
+
+        // invalid token should fail to log in
+        var client = createAuthenticatedClient(
+            env.store, [SyncCredentials.jwtIdToken(testInvalidJwtToken)]);
+
+        final events = <SyncLoginEvent>[];
+        client.loginEvents.listen(events.add);
+        client.start();
+        addTearDown(() => client.close());
+
+        expect(
+            await client.loginEvents.first.timeout(defaultTimeout,
+                onTimeout: () => throw TimeoutException(
+                    "Did not receive login event within $defaultTimeout")),
+            equals(SyncLoginEvent.credentialsRejected));
+
+        // valid token should succeed to log in
+        client.setCredentials(SyncCredentials.jwtIdToken(testJwtToken));
+
+        waitUntilLoggedIn(client);
+        await yieldExecution();
+
+        expect(
+            events,
+            equals(
+                [SyncLoginEvent.credentialsRejected, SyncLoginEvent.loggedIn]));
+      });
+    }, skip: "Test requires to manually run Sync server");
   } else {
     // TESTS to run when SYNC is NOT available
 
@@ -405,7 +528,7 @@ void main() {
 class SyncServer {
   Directory? _dir;
   int? _port;
-  Future<Process>? _process;
+  Process? _process;
 
   static bool isAvailable() {
     // Note: this causes an additional valgrind summary output with a leak.
@@ -423,31 +546,72 @@ class SyncServer {
     }
   }
 
-  Future<int> start({bool keepDb = false}) async {
-    _port ??= await _getUnusedPort();
+  static Future<File> _writeConfFile(
+      Directory directory, String contents) async {
+    if (!await directory.exists()) {
+      await directory.create(recursive: true);
+    }
 
-    _dir ??= Directory('testdata-sync-server-$_port');
+    final configFile = File('${directory.path}/sync-server-conf-test.json');
+    await configFile.writeAsString(contents); // overwrites by default
+
+    return configFile;
+  }
+
+  /// By default uses `--unsecured-no-authentication`. To use a custom
+  /// authentication configuration, pass [configContents].
+  /// It will be used to create a config file that is passed using `--conf`.
+  ///
+  /// Set [keepDb] to not delete an existing database before starting the
+  /// server.
+  Future<int> start({bool keepDb = false, String? configContents}) async {
+    final port = _port ??= await _getUnusedPort();
+    _port = port;
+
+    final dir = _dir ??= Directory('testdata-sync-server-$port');
+    _dir = dir;
     if (!keepDb) _deleteDb();
 
-    _process = Process.start('sync-server', [
-      '--unsecured-no-authentication',
-      '--db-directory=${_dir!.path}',
+    // Note: add arguments using the '--arg=value' syntax (unlike at the
+    // command line, which uses '--arg value')!
+    final arguments = [
+      '--db-directory=${dir.path}',
       '--model=${Directory.current.path}/test/objectbox-model.json',
-      '--bind=ws://127.0.0.1:$_port',
+      '--bind=ws://127.0.0.1:$port',
       '--admin-bind=http://127.0.0.1:${await _getUnusedPort()}'
-    ]);
+    ];
+    if (configContents != null && configContents.isNotEmpty) {
+      // Note: command line arguments overwrite values in a conf file
+      final configFile = await _writeConfFile(dir, configContents);
+      arguments.add("--conf=${configFile.absolute.path}");
+    } else {
+      arguments.add('--unsecured-no-authentication');
+    }
 
-    return _port!;
+    print("Starting Sync server with arguments: $arguments");
+    final process = await Process.start('sync-server', arguments);
+    _process = process;
+
+    // Make log output visible when running tests
+    stdout.addStream(process.stdout);
+    stderr.addStream(process.stderr);
+
+    return port;
+  }
+
+  /// Like [onlineAt], but assuming the [_port] of this server.
+  Future<void> online() async {
+    return onlineAt(_port!);
   }
 
   /// Wait for the server to respond to a simple http request.
   /// This simple check speeds up test by only trying to log in after the server
   /// has started, avoiding the reconnect backoff intervals altogether.
-  Future<void> online() async => Future(() async {
+  static Future<void> onlineAt(int port) async => Future(() async {
         final httpClient = HttpClient();
         while (true) {
           try {
-            await httpClient.get('127.0.0.1', _port!, '');
+            await httpClient.get('127.0.0.1', port, '');
             break;
           } on SocketException catch (e) {
             // Only retry if "Connection refused" (not using error codes as they
@@ -460,17 +624,17 @@ class SyncServer {
           }
         }
         httpClient.close(force: true);
-      }).timeout(defaultTimeout);
+      }).timeout(defaultTimeout,
+          onTimeout: () => throw TimeoutException(
+              "Server did not come online within $defaultTimeout ms"));
 
   Future<void> stop({bool keepDb = false}) async {
-    if (_process == null) return;
-    final proc = await _process!;
+    final proc = _process;
+    if (proc == null) return;
     _process = null;
     proc.kill(ProcessSignal.sigint);
     final exitCode = await proc.exitCode;
     if (exitCode != 0) {
-      await stdout.addStream(proc.stdout);
-      await stderr.addStream(proc.stderr);
       expect(await proc.exitCode, isZero);
     }
     if (!keepDb) _deleteDb();